services/api-gateway/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-jose[cryptography]>=3.3.0
aio-pika>=9.3.0
httpx>=0.24.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
redis>=5.0.1
prometheus-client>=0.17.1
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation-httpx>=0.41b0
